Move etcd conf test to etcddb_conf_test.go file
